chore(benchmark): moved benchmark and solve module to workflow subnamespace
