+++ tests/benchmark/workflow/__init__.py
@@ -1,0 +1,6 @@
+"""Main workflow steps of the benchmark."""
+
+from .benchmark import run_benchmark
+from .solve import time_solving
+
+__all__ = ['run_benchmark', 'time_solving']
